requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
pydantic[email]
motor==2.3.1
python-decouple==3.4
python-multipart==0.0.5
python-jose==3.2.0